+++ ta4j-core/src/main/java/org/ta4j/core/indicators/AbstractIndicator.java
@@ -27,16 +27,13 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.Indicator;
-import org.ta4j.core.num.Num;
 
 /**
  * Abstract {@link Indicator indicator}.
  */
 public abstract class AbstractIndicator<T> implements Indicator<T> {
 
-    /**
-     * The logger
-     */
+    /** The logger. */
     protected final transient Logger log = LoggerFactory.getLogger(getClass());
 
     private final BarSeries series;
@@ -58,11 +55,6 @@ public abstract class AbstractIndicator<T> implements Indicator<T> {
     @Override
     public String toString() {
         return getClass().getSimpleName();
-    }
-
-    @Override
-    public Num numOf(Number number) {
-        return series.numOf(number);
     }
 
 }