+++ ta4j-core/src/main/java/org/ta4j/core/indicators/RSIIndicator.java
@@ -51,13 +51,13 @@ public class RSIIndicator extends CachedIndicator<Num> {
         Num averageLoss = averageLossIndicator.getValue(index);
         if (averageLoss.isZero()) {
             if (averageGain.isZero()) {
-                return numOf(0);
+                return zero();
             } else {
-                return numOf(100);
+                return hundred();
             }
         }
         Num relativeStrength = averageGain.dividedBy(averageLoss);
         // compute relative strength index
-        return numOf(100).minus(numOf(100).dividedBy(numOf(1).plus(relativeStrength)));
+        return hundred().minus(hundred().dividedBy(one().plus(relativeStrength)));
     }
 }