+++ ta4j-core/src/main/java/org/ta4j/core/indicators/SMAIndicator.java
@@ -46,7 +46,7 @@ public class SMAIndicator extends CachedIndicator<Num> {
 
     @Override
     protected Num calculate(int index) {
-        Num sum = numOf(0);
+        Num sum = zero();
         for (int i = Math.max(0, index - barCount + 1); i <= index; i++) {
             sum = sum.plus(indicator.getValue(i));
         }