+++ ta4j-core/src/main/java/org/ta4j/core/indicators/adx/DXIndicator.java
@@ -47,10 +47,10 @@ public class DXIndicator extends CachedIndicator<Num> {
     protected Num calculate(int index) {
         Num pdiValue = plusDIIndicator.getValue(index);
         Num mdiValue = minusDIIndicator.getValue(index);
-        if (pdiValue.plus(mdiValue).equals(numOf(0))) {
-            return numOf(0);
+        if (pdiValue.plus(mdiValue).equals(zero())) {
+            return zero();
         }
-        return pdiValue.minus(mdiValue).abs().dividedBy(pdiValue.plus(mdiValue)).multipliedBy(numOf(100));
+        return pdiValue.minus(mdiValue).abs().dividedBy(pdiValue.plus(mdiValue)).multipliedBy(hundred());
     }
 
     @Override