+++ ta4j-core/src/main/java/org/ta4j/core/indicators/helpers/VolumeIndicator.java
@@ -46,7 +46,7 @@ public class VolumeIndicator extends CachedIndicator<Num> {
     @Override
     protected Num calculate(int index) {
         int startIndex = Math.max(0, index - barCount + 1);
-        Num sumOfVolume = numOf(0);
+        Num sumOfVolume = zero();
         for (int i = startIndex; i <= index; i++) {
             sumOfVolume = sumOfVolume.plus(getBarSeries().getBar(i).getVolume());
         }