+++ ta4j-core/src/main/java/org/ta4j/core/indicators/volume/ChaikinMoneyFlowIndicator.java
@@ -55,7 +55,7 @@ public class ChaikinMoneyFlowIndicator extends CachedIndicator<Num> {
     @Override
     protected Num calculate(int index) {
         int startIndex = Math.max(0, index - barCount + 1);
-        Num sumOfMoneyFlowVolume = numOf(0);
+        Num sumOfMoneyFlowVolume = zero();
         for (int i = startIndex; i <= index; i++) {
             sumOfMoneyFlowVolume = sumOfMoneyFlowVolume.plus(getMoneyFlowVolume(i));
         }