+++ ta4j-core/src/main/java/org/ta4j/core/num/DoubleNum.java
@@ -35,7 +35,7 @@ import java.util.function.Function;
  */
 public class DoubleNum implements Num {
 
-    private static final DoubleNum ZERO = DoubleNum.valueOf(0);
+    public static final DoubleNum ZERO = DoubleNum.valueOf(0);
     private static final DoubleNum ONE = DoubleNum.valueOf(1);
     private static final DoubleNum HUNDRED = DoubleNum.valueOf(100);